Introduce tempfile crate to make library unittest predictable
Temporary directories and files needed for tests were being
created with std::fs, which relies solely on the underlying
OS for the cleanup of those resources. Because of that, some
tests would fail because of conflicts in file creation.

The tempfile crate ensures that the temporary resources are
cleaned up at scope end.